third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/computed-style-page-break-inside-expected.txt

Test that page-break-inside property is not inherited

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(parent).pageBreakInside is "avoid"
PASS window.getComputedStyle(child).pageBreakInside is "auto"
PASS successfullyParsed is true

TEST COMPLETE
